# openapi.yaml
openapi: 3.0.0
info:
  title: Node.js Template API
  version: 1.0.0
paths:
  /users:
    post:
      summary: 사용자 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: 사용자 생성 성공
        '500':
          description: 서버 에러
    get:
      summary: 모든 사용자 조회
      responses:
        '200':
          description: 사용자 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    email:
                      type: string
        '500':
          description: 서버 에러
  /users/{id}:
    get:
      summary: 사용자 정보 조회
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 사용자 정보 조회 성공
        '404':
          description: 사용자를 찾을 수 없음
    put:
      summary: 사용자 정보 업데이트
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
      responses:
        '200':
          description: 사용자 정보 업데이트 성공
        '404':
          description: 사용자를 찾을 수 없음
    delete:
      summary: 사용자 삭제
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 사용자 삭제 성공
        '404':
          description: 사용자를 찾을 수 없음
components: {}
tags: []